src/main/java/org/mashti/jetson/lean/codec/ObjectCodec.java
/**
 * This file is part of jetson.
 *
 * jetson is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * jetson is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with jetson.  If not, see <http://www.gnu.org/licenses/>.
 */
package org.mashti.jetson.lean.codec;

import io.netty.buffer.ByteBuf;
import java.lang.reflect.Type;
import org.mashti.jetson.exception.RPCException;
import org.mashti.jetson.exception.UnknownTypeException;

/** @author Masih Hajiarabderkani (mh638@st-andrews.ac.uk) */
public class ObjectCodec implements Codec {

    @Override
    public boolean isSupported(final Type type) {

        return type != null && type instanceof Class<?> && Object.class.isAssignableFrom((Class<?>) type);
    }

    @Override
    public void encode(final Object value, final ByteBuf out, final Codecs codecs, final Type type) throws RPCException {

        if (value == null) {
            out.writeBoolean(true);
        }
        else {
            out.writeBoolean(false);
            final Class<?> runtime_type = value.getClass();
            if (runtime_type.equals(Object.class)) { throw new UnknownTypeException("type Object at runtime is not supported"); }
            codecs.encodeAs(runtime_type.getName(), out, String.class);
            codecs.encodeAs(value, out, value.getClass());
        }
    }

    @Override
    public Object decode(final ByteBuf in, final Codecs codecs, final Type type) throws RPCException {

        final boolean is_null = in.readBoolean();
        if (is_null) { return null; }

        final String value_type_as_string = codecs.decodeAs(in, String.class);
        final Class<?> value_type;
        try {
            value_type = Class.forName(value_type_as_string);
        }
        catch (final Exception e) {
            throw new RPCException(e);
        }
        return codecs.decodeAs(in, value_type);
    }
}
